Update index.html
The Bauhaus explored depth through movement. Right now, your Blue Block and Orange Circle are glued to the wall.

We should make them react to your mouse. When you move your cursor, the shapes should shift slightly in the opposite direction. This creates a "floating plane" effect, proving that the shapes are physical objects suspended in space.

This requires a tiny bit of JavaScript.